pyproject.toml
[tool.poetry]
name = "TC10281"
version = "0.1.0"
description = ""
authors = ["Sebastian Marines <sebastian0marines@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tk = "^0.1.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.910"
black = { version = "^21.9b0", allow-prereleases = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
